Make IntegratedFixture available from ff_mocks.py
