+++ test/conftest_settings.py
@@ -1,10 +1,11 @@
 import os
 
+from dcicutils.ff_mocks import IntegratedFixture
+
 
 TEST_DIR = os.path.join(os.path.dirname(__file__))
 
-# XXX: Refactor to config
-# INTEGRATED_ENV = 'fourfront-mastertest'
-INTEGRATED_ENV = 'fourfront-mastertest'
-INTEGRATED_ENV_INDEX_NAMESPACE = 'fourfront_mastertest'
-INTEGRATED_ENV_PORTAL_URL = 'https://mastertest.4dnucleome.org'
+INTEGRATED_ENV = IntegratedFixture.ENV_NAME
+INTEGRATED_ENV_INDEX_NAMESPACE = IntegratedFixture.ENV_INDEX_NAMESPACE
+INTEGRATED_ENV_PORTAL_URL = IntegratedFixture.ENV_PORTAL_URL
+INTEGRATED_ES = IntegratedFixture.ES_URL
